Move coral::util::TempDir to the public API
This fixes issue #9.
