+++ src/include/coral/util.hpp
@@ -190,59 +190,6 @@ ScopeGuard<Action> OnScopeExit(Action action) { return ScopeGuard<Action>(action
 
 
 /**
-\brief  An RAII object that creates a unique directory on construction
-        and recursively deletes it again on destruction.
-*/
-class TempDir
-{
-public:
-    /**
-    \brief  Creates a new temporary directory.
-
-    The name of the new directory will be randomly generated, and there are
-    three options of where it will be created, depending on the value of
-    `parent`.  In the following, `temp` refers to a directory suitable for
-    temporary files under the conventions of the operating system (e.g. `/tmp`
-    under UNIX-like systems), and `name` refers to the randomly generated
-    name mentioned above.
-
-      - If `parent` is empty: `temp/name`
-      - If `parent` is relative: `temp/parent/name`
-      - If `parent` is absolute: `parent/name`
-    */
-    explicit TempDir(
-        const boost::filesystem::path& parent = boost::filesystem::path());
-
-    TempDir(const TempDir&) = delete;
-    TempDir& operator=(const TempDir&) = delete;
-
-    /**
-    \brief  Move constructor.
-
-    Ownership of the directory is transferred from `other` to `this`.
-    Afterwards, `other` no longer refers to any directory, meaning that
-    `other.Path()` will return an empty path, and its destructor will not
-    perform any filesystem operations.
-    */
-    TempDir(TempDir&& other) CORAL_NOEXCEPT;
-
-    /// Move assignment operator. See TempDir(TempDir&&) for semantics.
-    TempDir& operator=(TempDir&&) CORAL_NOEXCEPT;
-
-    /// Destructor.  Recursively deletes the directory.
-    ~TempDir() CORAL_NOEXCEPT;
-
-    /// Returns the path to the directory.
-    const boost::filesystem::path& Path() const;
-
-private:
-    void DeleteNoexcept() CORAL_NOEXCEPT;
-
-    boost::filesystem::path m_path;
-};
-
-
-/**
 \brief  Starts a new process.
 
 Windows warning: This function only supports a very limited form of argument
